# lib/api-zod/src/generated/types/aiTestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiTestInput {
  provider: string;
  apiKey?: string;
  baseUrl?: string;
  model?: string;
  prompt?: string;
  useEnvKey?: boolean;
}