# lib/api-zod/src/generated/types/aiTestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiTestResult {
  success: boolean;
  message: string;
  response?: string;
  latency?: number;
  provider?: string;
  model?: string;
}