# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface AuthResult {
  success: boolean;
  student: Student;
  role: string;
}