# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: number;
  studentId: number;
  date: string;
  title: string;
  traz: number;
  rank: number;
  overallPercentage: number;
  /** @nullable */
  lessonsJson?: string | null;
  createdAt?: string;
}