# lib/api-zod/src/generated/types/providerResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderResult {
  success: boolean;
  message: string;
}