# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  code: string;
  field: string;
  grade: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  mainGoal?: string | null;
  /** @nullable */
  subscriptionType?: string | null;
  /** @nullable */
  paymentStatus?: string | null;
  /** @nullable */
  currentTraz?: number | null;
  /** @nullable */
  targetTraz?: number | null;
  /** @nullable */
  studyHoursPerDay?: number | null;
  createdAt?: string;
}