# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Taranom Mehr AI Education Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  code: string;
  field: string;
  grade: string;
  city?: string;
  age?: number;
  mainGoal?: string;
  subscriptionType?: string;
  currentTraz?: number;
  targetTraz?: number;
  studyHoursPerDay?: number;
}